requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine Learning & Optimization
tensorflow>=2.8.0
torch>=1.10.0
openvino>=2023.0.0
bayesian-optimization>=1.2.0
gym>=0.21.0
stable-baselines3>=1.5.0

# Simulation
simpy>=4.0.1
pyDOE2>=1.3.0
networkx>=2.6.0

# API & Web
fastapi>=0.75.0
uvicorn>=0.17.0
pydantic>=1.9.0
python-multipart>=0.0.5
python-dotenv>=0.20.0
email-validator>=1.3.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Intel Optimizations
intel-tensorflow>=2.8.0
intel-numpy>=1.20.0
daal4py>=2021.5.0
scikit-learn-intelex>=2021.5.0

# Development & Testing
pytest>=7.0.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0
httpx>=0.23.0
black>=22.1.0
isort>=5.10.0
flake8>=4.0.0
mypy>=1.3.0
